Add Zendesk\API\Traits\Utility\Pagination
