+++ src/Zendesk/API/Resources/Core/Tickets.php
@@ -12,7 +12,8 @@ use Zendesk\API\Traits\Resource\DeleteMany;
 use Zendesk\API\Traits\Resource\FindMany;
 use Zendesk\API\Traits\Resource\UpdateMany;
 use Zendesk\API\Traits\Utility\InstantiatorTrait;
-use Zendesk\API\Traits\Utility\CbpIterator;
+use Zendesk\API\Traits\Utility\Pagination\CbpStrategy;
+use Zendesk\API\Traits\Utility\Pagination\PaginationIterator;
 
 /**
  * The Tickets class exposes key methods for reading and updating ticket data
@@ -51,11 +52,13 @@ class Tickets extends ResourceAbstract
      *     process($ticket)
      * }
      *
-     * @return CbpIterator to fetch all pages.
+     * @return PaginationIterator to fetch all pages.
      */
     public function iterator()
     {
-        return new CbpIterator($this, 'tickets');
+        // TODO: default page size everywhere
+        $strategy = new CbpStrategy($this, 'tickets', 2);
+        return new PaginationIterator($strategy);
     }
 
     /**
