+++ src/Zendesk/API/Resources/Core/Users.php
@@ -14,7 +14,8 @@ use Zendesk\API\Traits\Resource\FindMany;
 use Zendesk\API\Traits\Resource\MultipartUpload;
 use Zendesk\API\Traits\Resource\UpdateMany;
 use Zendesk\API\Traits\Utility\InstantiatorTrait;
-use Zendesk\API\Traits\Utility\CbpIterator;
+use Zendesk\API\Traits\Utility\Pagination\CbpStrategy;
+use Zendesk\API\Traits\Utility\Pagination\PaginationIterator;
 
 /**
  * The Users class exposes user management methods
@@ -55,11 +56,12 @@ class Users extends ResourceAbstract
      *     process($user)
      * }
      *
-     * @return CbpIterator to fetch all pages.
+     * @return PaginationIterator to fetch all pages.
      */
     public function iterator()
     {
-        return new CbpIterator($this, 'users');
+        $strategy = new CbpStrategy($this, 'users', 2);
+        return new PaginationIterator($strategy);
     }
 
     /**
